SatCore.cc: バグ修正
* レベル0 で確定した変数の理由となっている節を削除しないようにした．

+++ c++-src/ymsat/core/SatCore.cc
@@ -474,7 +474,7 @@ SatCore::sweep_clause(
 	break;
       }
     }
-    if ( satisfied ) {
+    if ( satisfied && !is_locked(c) ) {
       // c を削除する．
       delete_clause(c);
     }
